URL Changing
Before the html pages were not loading the other files needed to work on
our server. I finally was able to figure out the correct url path for
our files. Instead of Image/india.gif it wants the server location in
front so for example:
https://www.secs.oakland.edu/~bdmullin/Image/1india.gif. I am still not
sure why the 1 is there. I did change the urls on home.html. We may need
to change all the pages and also the java plugin is not working.